test: 매칭 브랜치 충돌 해결
- test 환경 통합

ref:#72
